Added required version number
I was struggling to figure out why it wasn't working for me until Villhellm discord server pointed out that the version of `node-red-contrib-home-assistant-websocket` that came pre-installed with the HASSIO version of NodeRed doesn't work, and that I needed to upgrade it. Adding the minimum version number required to make it a bit more obvious for people to check for an update even if they already have it installed.